Add basic telegram support in all modules

+++ src/main/kotlin/api/objects/User.kt
@@ -1,13 +1,13 @@
 package api.objects
 
 abstract class BaseUser {
-        abstract val id: Int
+        abstract val id: Long
         abstract val name: String
         abstract val nick: String
 }
 
 data class VkUser(
-        override val id: Int,
+        override val id: Long,
         override val name: String,
         override val nick: String,  // aka `domain`
         val first_name: String,
@@ -17,5 +17,5 @@ data class VkUser(
         val bdate: String?,
         val online: Int?,
         val is_admin: Boolean,
-        val member_id: Int
+        val member_id: Long
         ) : BaseUser()
