+++ src/main/kotlin/modules/chatbot/chatBotEvents/LongPollEventBase.kt
@@ -4,7 +4,7 @@ import api.PlatformApiInterface
 
 abstract class LongPollEventBase {
     abstract val platform: Platform
-    abstract val chatId: Int
+    abstract val chatId: Long
     abstract val api: PlatformApiInterface
 }
